+++ src/test/resources/projects/allowdynamicimport/expected/dynamic-imports-extern.js
@@ -4,4 +4,4 @@ async function onMoreClicked(a) {
   a();
 }
 document.getElementById("loadMore").addEventListener(a => onMoreClicked(a));
-
+export{};